requirements.txt
python-telegram-bot==20.6
supabase==2.3.4
httpx==0.25.0
joblib==1.3.2
pandas==2.1.1
numpy==1.26.0
asyncio==3.4.3
tqdm==4.66.1
requests==2.31.0
